refactor: divide GameController per single feature and ViewController
